+++ autodocodec-api-usage/test/Autodocodec/Aeson/SchemaSpec.hs
@@ -78,6 +78,8 @@ spec = do
   jsonSchemaSpec @LegacyObject "legacy-object"
   jsonSchemaSpec @Ainur "ainur"
   jsonSchemaSpec @War "war"
+  jsonSchemaSpec @These "these"
+  jsonSchemaSpec @Expression "expression"
   describe "JSONSchema" $ do
     genValidSpec @JSONSchema
     xdescribe "does not hold because this property does not hold for Scientific values like -7.85483897507979979e17" $