+++ autodocodec-api-usage/test/Autodocodec/AesonSpec.hs
@@ -95,6 +95,8 @@ spec = do
   aesonCodecSpec @LegacyObject
   aesonCodecSpec @Ainur
   aesonCodecSpec @War
+  aesonCodecSpec @These
+  aesonCodecSpec @Expression
 
 aesonCodecErrorSpec ::
   forall a.