+++ autodocodec-openapi3/src/Autodocodec/OpenAPI/Schema.hs
@@ -17,6 +17,7 @@ import Control.Monad.State.Lazy (StateT, evalStateT, runStateT)
 import qualified Control.Monad.State.Lazy as State
 import Control.Monad.Trans (lift)
 import qualified Data.Aeson as Aeson
+import qualified Data.Foldable as Foldable
 import Data.HashMap.Strict (HashMap)
 import qualified Data.HashMap.Strict as HashMap
 import qualified Data.HashMap.Strict.InsOrd as InsOrdHashMap
@@ -182,6 +183,22 @@ declareNamedSchemaVia c' Proxy = evalStateT (go c') mempty
             u
             (NamedSchema Nothing (combineObjectSchemas s1s))
             (NamedSchema Nothing (combineObjectSchemas s2s))
+      DiscriminatedUnionCodec pn _ m -> do
+        let d =
+              Discriminator
+                { _discriminatorPropertyName = pn,
+                  _discriminatorMapping = InsOrdHashMap.fromHashMap $ fmap fst m
+                }
+            mkSchema dName (refName, oc) = do
+              s <- goObject $ oc *> (requiredFieldWith' pn (literalTextCodec dName) .= const dName)
+              declareSpecificSchemaRef (Just refName) $ combineObjectSchemas s
+        ss <- HashMap.traverseWithKey mkSchema m
+        pure
+          [ mempty
+              { _schemaDiscriminator = Just d,
+                _schemaOneOf = Just $ Foldable.toList ss
+              }
+          ]
       ApCodec oc1 oc2 -> do
         ss1 <- goObject oc1
         ss2 <- goObject oc2