+++ autodocodec-swagger2/src/Autodocodec/Swagger/Schema.hs
@@ -11,6 +11,8 @@ module Autodocodec.Swagger.Schema where
 
 import Autodocodec
 import Control.Monad
+import Data.Foldable (toList)
+import qualified Data.HashMap.Strict as HashMap
 import qualified Data.HashMap.Strict.InsOrd as InsOrdHashMap
 import Data.List
 import Data.Proxy
@@ -164,6 +166,14 @@ declareNamedSchemaVia c' Proxy = go c'
         ss1 <- goObject oc1
         ss2 <- goObject oc2
         pure [combineSchemaOr u (combineObjectSchemas ss1) (combineObjectSchemas ss2)]
+      DiscriminatedUnionCodec pn _ m -> do
+        let mkSchema dName (_, oc) =
+              fmap combineObjectSchemas $ goObject $ oc *> (requiredFieldWith' pn textCodec .= const dName)
+        ss <- HashMap.traverseWithKey mkSchema m
+        let combined = case toList ss of
+              [] -> mempty
+              (s : ss') -> foldr (combineSchemaOr DisjointUnion) s ss'
+        pure [combined]
       ApCodec oc1 oc2 -> do
         ss1 <- goObject oc1
         ss2 <- goObject oc2